add specialtype module; pass typecheck
